Callback fixes and improvements
Issue:
Callback 'timeout' of Net::Server::Mail is called when either read timeout exceeds or connection is closed by peer or an error occurs during reading from socket.

Solution:
Fixed event 'timeout'. The callback 'timeout' is only called when read timeout exceeds.
Added new callback 'stop_session' which is called when a connection is closed by peer or read error occurs. The callback takes one argument which is undefined in case when session is closed and contains error description in case of error.
Added documentation for the new callback.