Fix failing test, fixes GH #8:
As described in GH #8, there is a failing test which appears to
call `response` twice. The first time succeeds (and if `quit` is
called right after, that succeeds too).

The second time, however, called explicitly, it fails with a
time-out (since the server probably already responded and has
nothing more to respond). The following call to `quit` then fails
too.